requirements.txt
sentence-transformers==2.7.0
openai==1.35.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
rich==13.7.1
